Extract an interface for LoadingStrategy
